big file structure revamp. settings and globals got organized to help with dependency tree. fixes #44
