+++ src/BMC_Die.h
@@ -1,48 +1,22 @@
 ///////////////////////////////////////////////////////////////////////////////////////////
-// die.h
+// BMC_Die.h
 // Copyright (c) 2001-2024 Denis Papp. All rights reserved.
 // denis@accessdenied.net
 // https://github.com/pappde/bmai
-// 
-// DESC: main header for Die classes
 //
 // REVISION HISTORY:
 // drp030321 - partial split out to individual headers
+// dbl100524 - further split out of individual headers
 ///////////////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
 
-// TODO_HEADERS: drp030321 - cleanup headers
-#include "bmai.h"
+#include "BMC_BitArray.h"
+#include "BMC_DieData.h"
+#include "BMC_Move.h"
 
-// NOTES:
-// - second die only used for Option and Twin
-class BMC_DieData
-{
-	friend class BMC_Parser;
 
-public:
-	BMC_DieData();
-	virtual void Reset();
-
-	// methods
-	void		Debug(BME_DEBUG _cat = BME_DEBUG_ALWAYS);
-
-	// accessors
-	bool		HasProperty(INT _p) { return m_properties & _p; }
-	BME_SWING	GetSwingType(INT _d) { return (BME_SWING)m_swing_type[_d]; }
-	bool		Valid() { return (m_properties & BME_PROPERTY_VALID); }
-	// drp022521 - fixed to return INT instead of bool
-	INT			Dice() { return (m_properties & BME_PROPERTY_TWIN) ? 2 : 1; }
-	INT			GetSides(INT _t) { return m_sides[_t]; }
-
-protected:
-	U32			m_properties;
-	U8			m_sides[BMD_MAX_TWINS];			// number of sides if not a swing die, or current number of sides
-
-private:
-	U8			m_swing_type[BMD_MAX_TWINS];	// definition number of sides, should be BME_SWING
-};
+class BMC_Player;
 
 class BMC_Die : public BMC_DieData
 {
