+++ src/BMC_QAI.h
@@ -1,0 +1,28 @@
+///////////////////////////////////////////////////////////////////////////////////////////
+// BMC_QAI.h
+// Copyright (c) 2001-2024 Denis Papp. All rights reserved.
+// denis@accessdenied.net
+// https://github.com/pappde/bmai
+//
+// DESC:
+//
+// REVISION HISTORY:
+// drp030321 - partial split out to individual headers
+// dbl100524 - further split out of individual headers
+///////////////////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+#include "BMC_AI.h"
+
+
+// Heuristic Quick AI that BMAI uses by default
+class BMC_QAI : public BMC_AI
+{
+public:
+	virtual void		GetAttackAction(BMC_Game *_game, BMC_Move &_move);
+
+
+protected:
+private:
+};
