+++ src/BMC_Stats.cpp
@@ -1,5 +1,5 @@
 ///////////////////////////////////////////////////////////////////////////////////////////
-// stats.cpp
+// BMC_Stats.cpp
 // Copyright (c) 2001-2024 Denis Papp. All rights reserved.
 // denis@accessdenied.net
 // https://github.com/pappde/bmai
@@ -10,8 +10,9 @@
 // drp030321 - split out from mega source file
 ///////////////////////////////////////////////////////////////////////////////////////////
 
-#include "bmai.h"
-#include "stats.h"
+#include "BMC_Stats.h"
+
+#include <cstdio>
 
 ///////////////////////////////////////////////////////////////////////////////////////////
 // BMC_Stats
