--- src/types.h
@@ -1,26 +1,0 @@
-///////////////////////////////////////////////////////////////////////////////////////////
-// types.h
-// Copyright (c) 2001-2024 Denis Papp. All rights reserved.
-// denis@accessdenied.net
-// https://github.com/pappde/bmai
-// 
-// DESC: common types
-//
-// REVISION HISTORY:
-// drp060323 - split from bmai.h
-//
-///////////////////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-// typedefs
-typedef unsigned long	U32;
-typedef unsigned char	U8;
-typedef signed char		S8;
-typedef unsigned short	U16;
-typedef int				INT;
-typedef unsigned int	UINT;
-typedef float			F32;
-
-// special types
-typedef std::vector<float>	BMC_FloatVector;
